Factor out non-PySNMP-specific methods
A PySNMP-based and a netsnmp-cffi-based `TrapReceiver` implementation
will share a lot of common functionality that is unrelated to the
specifics of the underlying SNMP libraries.  This factors those things
out in to a base class that can be common to the implementations, to
avoid redundancies.
