+++ src/zino/trapd/pysnmp_backend.py
@@ -3,7 +3,7 @@
 import asyncio
 import logging
 from ipaddress import ip_address
-from typing import Dict, List, Optional
+from typing import Dict, Optional
 
 from pyasn1.type.base import SimpleAsn1Type
 from pysnmp.carrier.asyncio.dgram import udp
@@ -17,19 +17,17 @@ import zino.state
 from zino.config.models import PollDevice
 from zino.oid import OID
 from zino.snmp.pysnmp_backend import get_new_snmp_engine, mib_value_to_python
-from zino.statemodels import DeviceState, IPAddress
 from zino.trapd.base import (
     TrapMessage,
-    TrapObserver,
     TrapOriginator,
-    TrapType,
+    TrapReceiverBase,
     TrapVarBind,
 )
 
 _logger = logging.getLogger(__name__)
 
 
-class TrapReceiver:
+class TrapReceiver(TrapReceiverBase):
     """Zino Adapter for SNMP trap reception using PySNMP.
 
     A major difference to Zino 1 is that this receiver must explicitly be configured with SNMP community strings that
@@ -46,38 +44,8 @@ class TrapReceiver:
         state: Optional[zino.state.ZinoState] = None,
         polldevs: Optional[Dict[str, PollDevice]] = None,
     ):
-        self.transport: udp.UdpTransport = None
-        self.address = address
-        self.port = port
-        self.loop = loop if loop else asyncio.get_event_loop()
-        self.state = state or zino.state.ZinoState()
-        self.polldevs = polldevs if polldevs is not None else {}
+        super().__init__(address, port, loop, state, polldevs)
         self.snmp_engine = get_new_snmp_engine()
-        self._communities = set()
-        self._observers: dict[TrapType, List[TrapObserver]] = {}
-        self._auto_subscribed_observers = set()
-
-    def auto_subscribe_observers(self):
-        """Automatically subscribes all loaded TrapObserver subclasses to this trap receiver"""
-        for observer_class in TrapObserver.__subclasses__():
-            if not observer_class.WANTED_TRAPS:
-                continue
-            if observer_class in self._auto_subscribed_observers:
-                continue
-            else:
-                self._auto_subscribed_observers.add(observer_class)
-            observer_instance = observer_class(state=self.state, polldevs=self.polldevs, loop=self.loop)
-            self.observe(observer_instance, *observer_instance.WANTED_TRAPS)
-
-    def observe(self, subscriber: TrapObserver, *trap_types: List[TrapType]):
-        """Adds a trap subscriber to the receiver"""
-        for trap_type in trap_types:
-            observers = self._observers.setdefault(trap_type, [])
-            observers.append(subscriber)
-
-    def get_observers_for(self, trap_type: TrapType) -> List[TrapObserver]:
-        """Returns a list of trap observers for a given trap type"""
-        return self._observers.get(trap_type, [])
 
     async def open(self):
         """Opens the UDP transport socket and starts receiving traps"""
@@ -98,9 +66,7 @@ class TrapReceiver:
 
     def add_community(self, community: str):
         """Adds a new community string that will be accepted on incoming packets"""
-        if community in self._communities:
-            return
-        self._communities.add(community)
+        super().add_community(community)
         config.addV1System(self.snmp_engine, str(len(self._communities)), community)
 
     def trap_received(self, snmp_engine, state_reference, context_engine_id, context_name, var_binds, callback_context):
@@ -155,26 +121,6 @@ class TrapReceiver:
             return False
 
         return True
-
-    async def dispatch_trap(self, trap: TrapMessage):
-        """Dispatches incoming trap messages according to internal subscriptions"""
-        observers = self.get_observers_for((trap.mib, trap.name))
-        if not observers:
-            _logger.info("unknown trap: %s", trap)
-            return
-
-        for observer in observers:
-            try:
-                if not await observer.handle_trap(trap):
-                    return
-            except Exception:  # noqa
-                _logger.exception("Unhandled exception in trap observer %r", observer)
-
-    def _lookup_device(self, address: IPAddress) -> Optional[DeviceState]:
-        """Looks up a device from Zino's running state from an IP address"""
-        name = self.state.addresses.get(address)
-        if name in self.state.devices:
-            return self.state.devices[name]
 
     def _resolve_object_name(self, object_name: ObjectName) -> tuple[str, str, OID]:
         """Raises MibNotFoundError if oid in `object_name` can not be found"""
